Create simpler-media-status
simple media status script for polybar/i3blocks that will display "paused" if paused, the song & artist name if playing, and "shhh..." if no media players are found. 

works for Youtube music running in firefox on OpenSUSE Tumblweed, requires mpris-ctl and playerctl

Didn't know what license to use or what else to do, it's just a shell script and I'm really new to Github so if I messed something up and/or there's any issue just say the word and I'll fix it.